Prepare for 0.1 release
