+++ .cspell.json
@@ -4,6 +4,8 @@
   "words": [
     "Bitwarden",
     "chrono",
+    "Newtype",
+    "NOPAD",
     "Totp"
   ],
   "flagWords": [],
