+++ credential-exchange-protocol/README.md
@@ -1,7 +1,10 @@
-# Credential Exchange
+# Credential Exchange Protocol (CXP)
 
-Credential Exchange is a collection of Rust libraries for working with the
-[Credential Exchange Specifications](https://fidoalliance.org/specifications-credential-exchange-specifications/).
+This crate contains type definitions for the
+[FIDO Alliance’s credential exchange](https://fidoalliance.org/specifications-credential-exchange-specifications/)
+protocol specification. It's targeting the
+[Working Draft, October 03, 2024](https://fidoalliance.org/specs/cx/cxp-v1.0-wd-20241003.html)
+revision.
 
 For more information about the credential exchange protocol, please read the
 [Bitwarden blog post](https://bitwarden.com/blog/security-vendors-join-forces-to-make-passkeys-more-portable-for-everyone/)
@@ -20,11 +23,3 @@ or the
 <!-- prettier-ignore -->
 > [!NOTE]
 > This library is still in early development and as the specification evolves so will this library.
-
-## Structure
-
-It's currently comprised of the following crates:
-
-- `credential-exchange-format`: Credential Exchange Format.
-- `credential-exchange-protocol`: Credential Exchange Protocol.
-- `credential-exchange`: Reserved crate for higher level abstractions.
