--- credential-exchange-types/src/lib.rs
@@ -1,7 +1,0 @@
-mod b64url;
-pub use b64url::B64Url;
-
-pub mod format;
-pub mod protocol;
-
-type Uri = String;
