+++ credential-exchange/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "credential-exchange"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license-file.workspace = true
+keywords.workspace = true
