+++ credential-exchange/README.md
@@ -1,0 +1,15 @@
+# Credential Exchange
+
+This crate is a placeholder for reserving the name. For existing functionality please see
+[`credential-exchange-format`](https://crates.io/crates/credential-exchange-format) and
+[`credential-exchange-protocol`](https://crates.io/crates/credential-exchange-protocol).
+
+---
+
+Credential Exchange is a collection of Rust libraries for working with the
+[Credential Exchange Specifications](https://fidoalliance.org/specifications-credential-exchange-specifications/).
+
+For more information about the credential exchange protocol, please read the
+[Bitwarden blog post](https://bitwarden.com/blog/security-vendors-join-forces-to-make-passkeys-more-portable-for-everyone/)
+or the
+[Fido Alliance announcement](https://fidoalliance.org/fido-alliance-publishes-new-specifications-to-promote-user-choice-and-enhanced-ux-for-passkeys/).
